Activate Expect-Continue handshake for file upload
In certain cases when we try to upload a file the server can respond
back with a redirect and close the connection before we're done writing
content to the server which results in a 'java.net.SocketException:
Broken pipe' exception. We can solve that by asking the server first
whether we can actually write data (the rest is handled by the
httpclient library).

See also:
http://httpcomponents.10934.n7.nabble.com/Broken-pipe-Write-failed-when-making-Unauthorized-request-tp34235p34245.html

Signed-off-by: Luca Weiss <luca@z3ntu.xyz>