Increase wait time for wait_for_server
This script waits for the server to wake up during CI before launching tests.
In this library it only waited for 100 seconds instead of the 500 seconds it waits in the Files app.

This caused "failed to wait for server" in some occasions where the server image needs to be downloaded
and there's slow internet or poor performance.

Signed-off-by: Álvaro Brey <alvaro.brey@nextcloud.com>