CR: Handling client cert errors in NextcloudClient
During code review it was found that HTTP status 400 was only handled in `WebviewClient` based connections, but most connections are using `NextcloudClient`. So if a connection error occurs, most likely because of an expired certificate, it would not be handled correctly, and we would go on trying to use the expired certificate. So now a asic error handling was added.

However, so far I did not find a way to handle this kind of error in `OwnCloudClient` due to missing `Context`.

Signed-off-by: Elv1zz <elv1zz.git@gmail.com>