Update WebdavEntry.kt to handle images that are partially missing GPS metadata
This change to WebdavEntry.kt prevents a bug in the Android app. 

The bug occurs when the client loads a directory that contains an image that is partially (not completely) missing some GPS exif data. When the client attempts to load this directory, the app displays a "Server not available" message and fails to refresh.

Signed-off-by: Manuel Quinones <mquin2003@gmail.com>